requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
google-adk==0.3.0
google-cloud-aiplatform==1.57.0
httpx==0.27.0
asyncpg==0.29.0
python-dotenv==1.0.1
